+++ src/domain/alerts/alerts-decoders.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { DelayModifierDecoder } from '@/domain/alerts/contracts/delay-modifier-decoder.helper';
-import { MultiSendDecoder } from '@/domain/contracts/contracts/multi-send-decoder.helper';
-import { SafeDecoder } from '@/domain/contracts/contracts/safe-decoder.helper';
+import { DelayModifierDecoder } from '@/domain/alerts/contracts/decoders/delay-modifier-decoder.helper';
+import { MultiSendDecoder } from '@/domain/contracts/decoders/multi-send-decoder.helper';
+import { SafeDecoder } from '@/domain/contracts/decoders/safe-decoder.helper';
 
 @Module({
   providers: [DelayModifierDecoder, MultiSendDecoder, SafeDecoder],
