+++ src/domain/alerts/alerts.repository.ts
@@ -4,9 +4,9 @@ import { IAlertsRepository } from '@/domain/alerts/alerts.repository.interface';
 import { IAlertsApi } from '@/domain/interfaces/alerts-api.interface';
 import { AlertsRegistration } from '@/domain/alerts/entities/alerts.entity';
 import { AlertLog } from '@/routes/alerts/entities/alert.dto.entity';
-import { DelayModifierDecoder } from '@/domain/alerts/contracts/delay-modifier-decoder.helper';
-import { SafeDecoder } from '@/domain/contracts/contracts/safe-decoder.helper';
-import { MultiSendDecoder } from '@/domain/contracts/contracts/multi-send-decoder.helper';
+import { DelayModifierDecoder } from '@/domain/alerts/contracts/decoders/delay-modifier-decoder.helper';
+import { SafeDecoder } from '@/domain/contracts/decoders/safe-decoder.helper';
+import { MultiSendDecoder } from '@/domain/contracts/decoders/multi-send-decoder.helper';
 import { IEmailApi } from '@/domain/interfaces/email-api.interface';
 import { IAccountRepository } from '@/domain/account/account.repository.interface';
 import { ILoggingService, LoggingService } from '@/logging/logging.interface';
