+++ src/domain/contracts/encoders/multi-send-encoder.builder.ts
