+++ src/domain/contracts/encoders/safe-encoder.builder.ts
