+++ src/domain/relay/contracts/decoders/erc-20-decoder.helper.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { erc20Abi } from 'viem';
-import { AbiDecoder } from '@/domain/contracts/contracts/abi-decoder.helper';
+import { AbiDecoder } from '@/domain/contracts/decoders/abi-decoder.helper';
 
 @Injectable()
 export class Erc20Decoder extends AbiDecoder<typeof erc20Abi> {
