+++ src/domain/relay/contracts/decoders/proxy-factory-decoder.helper.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import ProxyFactory130 from '@/abis/safe/v1.3.0/GnosisSafeProxyFactory.abi';
-import { AbiDecoder } from '@/domain/contracts/contracts/abi-decoder.helper';
+import { AbiDecoder } from '@/domain/contracts/decoders/abi-decoder.helper';
 
 @Injectable()
 export class ProxyFactoryDecoder extends AbiDecoder<typeof ProxyFactory130> {
